rename centralized to single
